Beginnings of the poly_query
This is super messy. Just wanted to get the ability to draw a polygon on the map and have the flowtools return all the catchments in the polygon. The leaflet map is garage, but it works as a POC. Added a poly_query.py file. Did NOT create a class in this like in splitcatchment and flowtrace. This is very barebones and needs to be cleaned up.