Edits to the get_local_flowlines function
Figured out how to get the outlet flowlines for the polygon query. Still trying to figure out how to get the distance downstream querys to work for the outline flowlines. The first query is to the geoserver, the second is to nldi. The thought right now is to do these API calls, make them shapely multilinestrings, then merge them after.